HZN-1495: get jaeger-tracing to work with osgi
