+++ core/ipc/rpc/api/pom.xml
@@ -26,16 +26,4 @@
       </plugin>
     </plugins>
   </build>
-  <dependencies>
-    <dependency>
-      <groupId>io.opentracing</groupId>
-      <artifactId>opentracing-api</artifactId>
-      <version>${opentracingVersion}</version>
-    </dependency>
-    <dependency>
-      <groupId>io.opentracing</groupId>
-      <artifactId>opentracing-util</artifactId>
-      <version>${opentracingVersion}</version>
-    </dependency>
-  </dependencies>
 </project>
