--- core/ipc/rpc/jaeger-tracer/src/main/java/org/opennms/core/ipc/rpc/jaeger/tracer/JaegerTracerWrapper.java
@@ -1,60 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2019 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2019 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.core.ipc.rpc.jaeger.tracer;
-
-import org.opennms.core.rpc.api.TracerWrapper;
-import org.opennms.core.sysprops.SystemProperties;
-
-import io.jaegertracing.Configuration;
-import io.jaegertracing.internal.samplers.ConstSampler;
-import io.opentracing.Tracer;
-import io.opentracing.util.GlobalTracer;
-
-public class JaegerTracerWrapper implements TracerWrapper {
-
-    private static final String JAEGER_SAMPLER_PARAM_PROPERTY = "org.opennms.rpc.tracer.jaeger.sampler.param";
-    private static final int JAEGER_SAMPLER_PARAM_VALUE = SystemProperties.getInteger(JAEGER_SAMPLER_PARAM_PROPERTY, 1);
-    
-    @Override
-    public Tracer init(String serviceName) {
-        Configuration.SamplerConfiguration samplerConfig = Configuration.SamplerConfiguration.fromEnv()
-                .withType(ConstSampler.TYPE)
-                .withParam(JAEGER_SAMPLER_PARAM_VALUE);
-
-        Configuration.ReporterConfiguration reporterConfig = Configuration.ReporterConfiguration.fromEnv()
-                .withLogSpans(true);
-
-        Configuration config = new Configuration(serviceName)
-                .withSampler(samplerConfig)
-                .withReporter(reporterConfig);
-        GlobalTracer.get();
-
-        return config.getTracer();
-    }
-}
