--- core/ipc/rpc/jaeger-tracer/src/main/resources/META-INF/opennms/applicationContext-rpc-jaeger-tracer.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-       xmlns:tx="http://www.springframework.org/schema/tx"
-       xmlns:context="http://www.springframework.org/schema/context"
-       xmlns:util="http://www.springframework.org/schema/util"
-       xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
-       xsi:schemaLocation="
-  http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
-  http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-  http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-4.2.xsd
-  http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-4.2.xsd
-  http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
-  http://camel.apache.org/schema/spring http://camel.apache.org/schema/spring/camel-spring-2.19.1.xsd
-">
-
-  <context:annotation-config />
-
-  <bean id="jeagerTracerWrapper" class="org.opennms.core.ipc.rpc.jaeger.tracer.JaegerTracerWrapper" />
-
-  <onmsgi:service ref="jeagerTracerWrapper" interface="org.opennms.core.rpc.api.TracerWrapper" />
-
-</beans>
