--- core/ipc/rpc/jaeger-tracer/src/main/resources/OSGI-INF/blueprint/blueprint.xml
@@ -1,17 +1,0 @@
-<blueprint xmlns="http://www.osgi.org/xmlns/blueprint/v1.0.0"
-           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:cm="http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0"
-           xmlns:ext="http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0"
-           xsi:schemaLocation="
-		http://www.osgi.org/xmlns/blueprint/v1.0.0
-		https://osgi.org/xmlns/blueprint/v1.0.0/blueprint.xsd
-		http://aries.apache.org/blueprint/xmlns/blueprint-cm/v1.3.0
-		http://aries.apache.org/schemas/blueprint-cm/blueprint-cm-1.3.0.xsd
-		http://aries.apache.org/blueprint/xmlns/blueprint-ext/v1.5.0
-		http://aries.apache.org/schemas/blueprint-ext/blueprint-ext-1.5.xsd
-">
-
-  <bean id="jaegerTracerWrapper" class="org.opennms.core.ipc.rpc.jaeger.tracer.JaegerTracerWrapper" />
-
-  <service interface="org.opennms.core.rpc.api.TracerWrapper" ref="jaegerTracerWrapper" />
-
-</blueprint>
