+++ core/ipc/rpc/kafka/pom.xml
@@ -70,6 +70,11 @@
         <artifactId>rate-limited-logger</artifactId>
         <version>${rateLimitedLoggerVersion}</version>
       </dependency>
+      <dependency>
+        <groupId>org.opennms.core.tracing</groupId>
+        <artifactId>org.opennms.core.tracing.api</artifactId>
+        <version>${project.version}</version>
+      </dependency>
      <dependency>
         <groupId>org.opennms.core.test-api</groupId>
         <artifactId>org.opennms.core.test-api.lib</artifactId>
