+++ core/ipc/rpc/kafka/src/main/java/org/opennms/core/ipc/rpc/kafka/KafkaRpcServerManager.java
@@ -65,7 +65,7 @@ import org.opennms.core.ipc.rpc.kafka.tracing.RequestCarrier;
 import org.opennms.core.rpc.api.RpcModule;
 import org.opennms.core.rpc.api.RpcRequest;
 import org.opennms.core.rpc.api.RpcResponse;
-import org.opennms.core.rpc.api.TracerRegistry;
+import org.opennms.core.tracing.api.TracerRegistry;
 import org.opennms.distributed.core.api.MinionIdentity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -148,7 +148,7 @@ public class KafkaRpcServerManager {
                 }
             }
         });
-        tracer = tracerRegistry.getTracer(minionIdentity.getId()+"@"+minionIdentity.getLocation());
+        tracer = getTracer();
     }
 
     @SuppressWarnings({ "rawtypes", "unchecked" })
@@ -273,15 +273,22 @@ public class KafkaRpcServerManager {
                             }
                             Map<String, String> tracingInfoMap = RequestCarrier.getTracingInfoMap(rpcMessage.getTracingInfoList());
                             Tracer.SpanBuilder spanBuilder;
+                            if(tracerRegistry.isRegistered()) {
+                                tracer = getTracer();
+                            }
                             SpanContext context = tracer.extract(Format.Builtin.TEXT_MAP, new TextMapExtractAdapter(tracingInfoMap));
                             if (context != null) {
                                 spanBuilder = tracer.buildSpan(module.getId()).asChildOf(context);
                             } else {
                                 spanBuilder = tracer.buildSpan(module.getId());
                             }
+
                             Span minionSpan = spanBuilder.start();
-                            minionSpan.setTag("rpcId", rpcId);
                             RpcRequest request = module.unmarshalRequest(rpcContent.toStringUtf8());
+                            minionSpan.setTag("rpcId", rpcId);
+                            if(request.getSystemId() != null) {
+                                minionSpan.setTag("systemId", request.getSystemId());
+                            }
                             CompletableFuture<RpcResponse> future = module.execute(request);
                             future.whenComplete((res, ex) -> {
                                 final RpcResponse response;
@@ -294,6 +301,7 @@ public class KafkaRpcServerManager {
                                     // No exception occurred, use the given response
                                     response = res;
                                 }
+
                                 minionSpan.finish();
                                 try {
                                     final JmsQueueNameFactory topicNameFactory = new JmsQueueNameFactory(KafkaRpcConstants.RPC_RESPONSE_TOPIC_NAME,
@@ -346,6 +354,9 @@ public class KafkaRpcServerManager {
 
     }
 
+    private Tracer getTracer() {
+        return tracerRegistry.getTracer(minionIdentity.getId()+"@"+minionIdentity.getLocation());
+    }
     /**
      * RpcId is used to remove rpcId from DelayQueue after it reaches expirationTime.
      */
