+++ core/tracing/jaeger-osgi/pom.xml
@@ -1,0 +1,105 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <artifactId>org.opennms.core.tracing</artifactId>
+    <groupId>org.opennms.core</groupId>
+    <version>24.0.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.opennms.core.tracing</groupId>
+  <artifactId>org.opennms.core.tracing.jaeger-osgi</artifactId>
+  <packaging>bundle</packaging>
+  <name>OpenNMS :: Core :: Tracing :: Jaeger :: OSGi wrapper</name>
+  <description>OSGi wrapper for Jaeger Tracer.</description>
+
+  <properties>
+    <apacheThriftVersion>0.12.0</apacheThriftVersion>
+  </properties>
+  <build>
+    <defaultGoal>install</defaultGoal>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-shade-plugin</artifactId>
+        <version>2.4.2</version>
+        <executions>
+          <execution>
+            <phase>package</phase>
+            <goals>
+              <goal>shade</goal>
+            </goals>
+            <configuration>
+              <artifactSet>
+                <excludes>
+                  <exclude>io.opentracing:opentracing-api</exclude>
+                  <exclude>io.opentracing:opentracing-util</exclude>
+                  <exclude>io.opentracing:opentracing-noop</exclude>
+                  <exclude>io.opentracing.contrib:opentracing-tracerresolver</exclude>
+                  <exclude>io.jaegertracing:jaeger-tracerresolver</exclude>
+                  <exclude>com.google.code.gson:gson</exclude>
+                  <exclude>org.apache.httpcomponents:httpclient</exclude>
+                  <exclude>org.apache.httpcomponents:httpcore</exclude>
+                  <exclude>commons-codec:commons-codec</exclude>
+                  <exclude>com.squareup.okhttp3:okhttp</exclude>
+                  <exclude>com.squareup.okio:okio</exclude>
+                </excludes>
+              </artifactSet>
+              <filters>
+                <filter>
+                  <artifact>*:*</artifact>
+                </filter>
+              </filters>
+              <createDependencyReducedPom>true</createDependencyReducedPom>
+            </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <groupId>org.apache.felix</groupId>
+        <artifactId>maven-bundle-plugin</artifactId>
+        <extensions>true</extensions>
+        <configuration>
+          <instructions>
+            <Import-Package>
+              io.opentracing.util,
+              io.opentracing,
+              io.opentracing.propagation,
+              io.opentracing.tag
+            </Import-Package>
+            <Export-Package>
+              io.jaegertracing,
+              io.jaegertracing.internal,
+              io.jaegertracing.internal.reporters.*
+            </Export-Package>
+          </instructions>
+          <unpackBundle>true</unpackBundle>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+
+  <dependencies>
+    <dependency>
+      <groupId>io.jaegertracing</groupId>
+      <artifactId>jaeger-client</artifactId>
+      <version>${jeagertracingVersion}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.thrift</groupId>
+      <artifactId>libthrift</artifactId>
+      <version>${apacheThriftVersion}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.jaegertracing</groupId>
+      <artifactId>jaeger-core</artifactId>
+      <version>${jeagertracingVersion}</version>
+    </dependency>
+    <dependency>
+      <groupId>io.jaegertracing</groupId>
+      <artifactId>jaeger-thrift</artifactId>
+      <version>${jeagertracingVersion}</version>
+    </dependency>
+  </dependencies>
+</project>
