+++ core/tracing/jaeger-tracer/src/main/java/org/opennms/core/tracing/jaeger/ConditionalJaegerTracerContext.java
@@ -1,0 +1,64 @@
+/*******************************************************************************
+ * This file is part of OpenNMS(R).
+ *
+ * Copyright (C) 2019 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2019 The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
+ *
+ * OpenNMS(R) is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ *
+ * OpenNMS(R) is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with OpenNMS(R).  If not, see:
+ *      http://www.gnu.org/licenses/
+ *
+ * For more information contact:
+ *     OpenNMS(R) Licensing <license@opennms.org>
+ *     http://www.opennms.org/
+ *     http://www.opennms.com/
+ *******************************************************************************/
+
+package org.opennms.core.tracing.jaeger;
+
+import org.springframework.context.annotation.ConditionContext;
+import org.springframework.context.annotation.Conditional;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.ConfigurationCondition;
+import org.springframework.context.annotation.ImportResource;
+import org.springframework.core.type.AnnotatedTypeMetadata;
+
+/**
+ * Conditionally load Jaeger tracer if jaeger tracer is enabled in system properties.
+ */
+@Configuration
+@Conditional(ConditionalJaegerTracerContext.Condition.class)
+@ImportResource("/META-INF/opennms/applicationContext-tracer-jaeger.xml")
+public class ConditionalJaegerTracerContext {
+
+    public static final String TRACER = "org.opennms.core.tracer";
+    private static final String JEAGER_TRACER = "jaeger";
+    private static final String GLOBAL_TRACER = "global";
+
+    static class Condition implements ConfigurationCondition {
+
+        @Override
+        public ConfigurationPhase getConfigurationPhase() {
+            return ConfigurationPhase.PARSE_CONFIGURATION;
+        }
+
+        @Override
+        public boolean matches(ConditionContext context, AnnotatedTypeMetadata metadata) {
+            String tracerProperty = System.getProperty(TRACER, GLOBAL_TRACER);
+            final boolean enabled = JEAGER_TRACER.equals(tracerProperty);
+            return enabled;
+        }
+    }
+}
