+++ core/tracing/pom.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <parent>
+    <groupId>org.opennms</groupId>
+    <artifactId>org.opennms.core</artifactId>
+    <version>24.0.0-SNAPSHOT</version>
+  </parent>
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>org.opennms.core</groupId>
+  <artifactId>org.opennms.core.tracing</artifactId>
+  <packaging>pom</packaging>
+  <name>OpenNMS :: Core :: Tracing</name>
+  <modules>
+    <module>api</module>
+    <module>registry</module>
+    <module>jaeger-tracer</module>
+    <module>jaeger-osgi</module>
+  </modules>
+</project>
