+++ features/minion/repository/pom.xml
@@ -59,7 +59,8 @@
                                 <feature>minion-icmp-proxy</feature>
                                 <feature>minion-telemetryd-receivers</feature>
                                 <feature>hawtio-offline</feature>
-				<feature>cxf-commands/${cxfVersion}</feature>
+				                <feature>cxf-commands/${cxfVersion}</feature>
+                                <feature>jeager-tracer</feature>
                             </features>
                             <repository>${project.build.directory}/maven-repo</repository>
                         </configuration>
